requirements.txt
aniso8601==3.0.0
argcomplete==1.9.3
asn1crypto==0.24.0
azure-common==1.1.8
azure-nspkg==2.0.0
azure-storage==0.36.0
base58==0.2.4
boto3==1.7.1
botocore==1.10.1
certifi==2018.1.18
cffi==1.11.5
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
cognitive-face==1.4.1
configparser==3.5.0
cryptography==2.3
docutils==0.14
durationpy==0.5
Flask==0.12.2
Flask-RESTful==0.3.6
future==0.16.0
hjson==3.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.19.0
MarkupSafe==1.0
Pillow==5.0.0
placebo==0.8.1
pycparser==2.18
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2018.3
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
toml==0.9.4
tqdm==4.19.1
troposphere==2.2.1
Unidecode==1.0.22
urllib3==1.22
Werkzeug==0.13
wsgi-request-logger==0.4.6
zappa==0.45.1
